requirements.txt
# Direct dependencies for catpair utilities
Pillow>=10.0.0
pillow-heif>=0.13.0
